feat #169 : 챌린지 생성 결과 화면 이동

+++ android/data/src/main/java/com/sixkids/data/api/ChallengeService.kt
@@ -16,7 +16,7 @@ interface ChallengeService {
     @POST("challenges")
     suspend fun createChallenge(
         @Body challengeCreateRequest: ChallengeCreateRequest
-    ): ApiResult<ApiResponse<Long>>
+    ): ApiResult<ApiResponse<Int>>
 
     @GET("challenges")
     suspend fun getChallengeHistory(
